requirements.txt
streamlit>=1.36
python-pptx>=0.6.23
openai>=1.40
requests>=2.32
